requirements.txt
fastapi
uvicorn
google-generativeai
pydantic
python-multipart
docx
pdfplumber
